Web: [TEST] for create usecase (kernel build and reboot)
This commit adds units tests for the "Create testruns" use cases that
are already live - kernel_reboot and kernel_build. It uses
WWW::Mechanize. The test expects its use case files in a separate path
that was added to the configuration file. The use case files where moved
into this path.