[RFK] Carved out installer grub file
The Tapper-MCP had a hardcoded grub entry to
be used by the machine installer. This grub entry
is now contained in the default config file.

The Tapper-MCP still contains a hardcoded
fallback for backcompat reasons, so the entry
in tapper.yml wouldn't be strongly required,
but we use this opportunity to test subtle
YAML multiline and indented whitespace ability
for config entries.