defaults for backwards compatibility
During deployment no existing /etc/tapper.cfg
is overwritten. But we now depend on new values
in Tapper::Config, so we add them globally to
values that correspond to the former hardcoded
text in several Perl modules (::MCP and ::Installer).

Local config in /etc/tapper.cfg is used when available.